fix miscellaneous `numpy=2.0` errors (#8117)
* replace `np.unicode_` with `np.str_`

* replace `np.NaN` with `np.nan`

* replace more instances of `np.unicode_`

note that with more modern versions of `numpy` the `.astype(np.str_)`
don't actually change the dtype, so maybe we can remove those.

* more instances of renamed / removed dtypes

* more dtype replacements

* use `str.encode(encoding)` instead of `bytes(str, encoding)`

* explicitly import `RankWarning`

* left-over `np.RankWarning`

* use `float` instead of the removed `np.float_`

* ignore missing stubs for `numpy.exceptions`

---------

+++ xarray/core/missing.py
@@ -730,7 +730,7 @@ def interp_func(var, x, new_x, method: InterpOptions, kwargs):
         # scipy.interpolate.interp1d always forces to float.
         # Use the same check for blockwise as well:
         if not issubclass(var.dtype.type, np.inexact):
-            dtype = np.float_
+            dtype = float
         else:
             dtype = var.dtype
